adk-multiagent-example/requirements.txt
# Google ADK Agent Requirements
google-adk>=1.12.0
python-dotenv>=1.0.0

# Evaluation & Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
